renamed "supervision" to "supervized"
