+++ synchronization/supervized/supervisor.go
@@ -1,4 +1,4 @@
-package supervision
+package supervized
 
 import (
 	"context"
