+++ synchronization/supervized/supervisor_test.go
@@ -1,4 +1,4 @@
-package supervision
+package supervized
 
 import (
 	"context"
